api changes

+++ src/ISeeYou/Domain/Aggregates/Subject/Commands/AddPhotoLike.cs
@@ -8,7 +8,7 @@ namespace ISeeYou.Domain.Aggregates.Subject.Commands
         public int SubjectId { get; set; }
         public DateTime? StartDate { get; set; }
         public DateTime EndDate { get; set; }
-        public int PhotoId { get; set; }
+        public long PhotoId { get; set; }
         public int SourceId { get; set; }
         public string Image { get; set; }
         public string ImageBig { get; set; }
