+++ src/ISeeYou/Domain/Aggregates/Subject/Events/PhotoLikeAdded.cs
@@ -8,7 +8,7 @@ namespace ISeeYou.Domain.Aggregates.Subject
         public int SubjectId { get; set; }
         public DateTime? StartDate { get; set; }
         public DateTime EndDate { get; set; }
-        public int PhotoId { get; set; }
+        public long PhotoId { get; set; }
         public int SourceId { get; set; }
         public string Image { get; set; }
     }
